+++ GrailDiary/BookImporter/ImportForm.swift
@@ -5,10 +5,11 @@ import UniformTypeIdentifiers
 
 /// A form that lets the user input parameters for a book import job
 struct ImportForm: View {
-  var importAction: ([URL], Bool, Bool) -> Void
+  var importAction: ([URL], String, Bool, Bool) -> Void
   @State private var downloadCoverImages = false
   @State private var dryRun = true
   @State private var showDocumentPicker = false
+  @State private var hashtags = ""
 
   var body: some View {
     NavigationView {
@@ -16,6 +17,9 @@ struct ImportForm: View {
         Section(header: Text("Download")) {
           Toggle("Cover Images", isOn: $downloadCoverImages)
           Toggle("Dry Run", isOn: $dryRun)
+        }
+        Section(header: Text("Hashtags (Optional)"), footer: Text("Enter #hashtags for all imported books. Example: #goodreads")) {
+          TextField("#hashtag", text: $hashtags)
         }
         Button("Select File") {
           showDocumentPicker = true
@@ -25,8 +29,8 @@ struct ImportForm: View {
     }
     .navigationViewStyle(StackNavigationViewStyle())
     .sheet(isPresented: $showDocumentPicker, content: {
-      DocumentPickerView(contentTypes: [.json]) { urls in
-        importAction(urls, downloadCoverImages, dryRun)
+      DocumentPickerView(contentTypes: [.json, .commaSeparatedText]) { urls in
+        importAction(urls, hashtags, downloadCoverImages, dryRun)
       }
     })
   }
@@ -34,7 +38,7 @@ struct ImportForm: View {
 
 struct ImportForm_Previews: PreviewProvider {
   static var previews: some View {
-    ImportForm(importAction: { _, _, _ in })
+    ImportForm(importAction: { _, _, _, _ in })
 //      .previewDevice("iPod touch (7th generation)")
   }
 }